Added acceleration to tanks.
TANK_SPEED in constants now represent the players max_speed, TANK_ACCELERATION represents the players acceleraton.
players moving variable is set to false every frame in the end of players update function, but set back to true in the keypress functions(which are active if the player moves).
If the moving variable if false in the beggining of the update function (if no keypress function sets it to true), the players speed is set to 0 for new acceleration.