Fixed the error that said integer was passed to keymap dictionary.
Error was in controller where previous merge was resolved by leaving both the old and the new for-loops looping through the controllers keymap dictionary.
Removed the old loops --> Sucess.
If going back to handling every single event with pygame.event.get() instead of pygame.key.get_pressed() and event.get(), the commit before this would be a wise place to copy the old loops from.