Added slider class to the menu file in order to create the settings audio section. You can now control the volumes of all channels and the master volume from the settings state.
Visually the slider class contains a background image, knob image and caption. Rest works quite similar to the button class, except for an extra event - MOUSEBUTTONUP wich controls if you are dragging the slider or not.
Also added all the default volumes to constants since they belong there, and they need to be there in order to spawn the knobs on the right place when opening settings.