Big improvements for collision-detection. Also more effective initiation of WorldObjects.
Collision detection now properly uses different methods to detect collision for different types of WorldObjects.
Right now the two types of objects are:
* 0 = Normal Object (Bush, DeadBush etc)
* 1 = Area Object (Water)

Finally, many changes have been made to how map.py works internally which have been tested to still work well with other parts of the game.

The code used for collision-detection between player and object could in the future be modified to work with player and player and bullet(ammo) and player.