Many small changes and fixes in various parts of the project.
Minor improvements in the collision-handling in agent.py.Objects will now use a .solid - attribute that determines if and how a player can drive through it.

Fixed a small bug in ammo.py where the X-max was used instead of Y-max when deleting bullets outside of screen.
Added a temporary fix in controller.py to not crash the game when one of the players die.

Added some docstrings to describe the purposes of for example different classes and their methods. Feel free to edit them if you want since what I've written so far just is a start.
Also some cleaning up in the code.