Added borders to the map to limit players movement to the game world.
Also a small improvement on solid WorldObjects pushback to get smoother movements.
Tried to decrease speed of players using the .solid-variable from for example water-objects without success. Since the code isn't causing any serious bugs that I know of, I kept it in the commit to inspire someone to a solution that actually works.