Big improvements on the games performance.
Since the world only is changing when a WorldObject gets destroyed we only need to draw an updated world when that occurs. Because of this, the world will now draw a "cached" surface that represents the world instead of drawing every individual object.
This cached surface only gets updated when a flag-variable is set to True, which at the moment only occurs when a WorldObject with health gets destroyed by Ammo-objects.

Also minor changes to players default rotation